requirements.txt
Package           Version
----------------- -------
astroid           2.12.12
click             8.1.3
dill              0.3.6
Flask             2.2.2
Flask-MySQL       1.5.2
Flask-MySQLdb     1.0.1
isort             5.10.1
itsdangerous      2.1.2
jedi              0.18.1
Jinja2            3.1.2
lazy-object-proxy 1.8.0
MarkupSafe        2.1.1
mccabe            0.7.0
mysqlclient       2.1.1
parso             0.8.3
pip               22.0.2
platformdirs      2.5.3
PyJWT             2.6.0
pylint            2.15.5
PyMySQL           1.0.2
python-dotenv     0.21.0
setuptools        59.6.0
tomli             2.0.1
tomlkit           0.11.6
Werkzeug          2.2.2
wrapt             1.14.1
